Move row group pruning logic into pruning.rs
